clean: move handleDrag tests in a separate file
